feat: 괄호 (, ), 캐럿 ^ 처리 로직을 추가합니다.
-커스텀 구분자 안에 괄호나 캐럿이 있는지 판단합니다.
-괄호나 캐럿이 있다면 해당 기호 앞에 백슬래시 '\\'를 추가합니다.
-기능 구현 목록을 갱신합니다.
