+++ README.md
@@ -245,7 +245,7 @@ More : [Commit Message convention](https://gist.github.com/9941e89d80e2bc58a153.
     - [x] 구한 덧셈 결과를 컴퓨터에게 반환한다.
     - [ ] 특수 기호를 구분자로 사용한다.
        - [x] 백슬래시 `\`를 구분자로 사용할 경우 해당 기호 뒤에 `\\`를 추가한다.
-       - [ ] 괄호 `(, )`나 캐럿 `^`을 구분자로 사용할 경우 해당 기호 앞에 `\\`를 추가한다.
+       - [x] 괄호 `(, )`나 캐럿 `^`을 구분자로 사용할 경우 해당 기호 앞에 `\\`를 추가한다.
 
 ### 입출력 구현 리스트
