+++ src/main/java/calculator/util/CalculateCustom.java
@@ -8,8 +8,6 @@ public class CalculateCustom extends CalculateSeparator {
     private static final String CUSTOM_PRE = "//";
     private static final String CUSTOM_POST = "\\n";
     private static final Integer TWO = 2;
-    private static final Character BACK_SLASH = '\\';
-    private static final String STRING_BACK_SLASH = "\\";
 
     public static boolean checkNotContainCustom(String userInput) {
         return checkNotContainSeparator(userInput, CUSTOM_PRE) || checkNotContainSeparator(userInput, CUSTOM_POST);
@@ -42,28 +40,9 @@ public class CalculateCustom extends CalculateSeparator {
 
     public static List<String> splitWithCustomSeparator(String userInput) {
         String custom = getCustomSeparator(userInput);
-        if (checkContainBackSlash(custom)) {
-            return splitWithSeparator(getUserInput(userInput), createNewCustom(custom));
+        if (CalculateSpecialSeparator.checkContainBackSlash(custom) || CalculateSpecialSeparator.checkContainSpecial(custom)) {
+            return splitWithSeparator(getUserInput(userInput), CalculateSpecialSeparator.createNewCustom(custom));
         }
         return splitWithSeparator(getUserInput(userInput), getCustomSeparator(userInput));
-    }
-
-    public static boolean checkContainBackSlash(String custom) {
-        return custom.contains(STRING_BACK_SLASH);
-    }
-
-    public static String createNewCustom(String custom) {
-        StringBuilder newCustom = new StringBuilder();
-        for (int i = 0; i < custom.length(); i++) {
-            newCustom.append(custom.charAt(i));
-            if (checkIsCustomEqualBackSlash(custom.charAt(i))) {
-                newCustom.append(BACK_SLASH);
-            }
-        }
-        return newCustom.toString();
-    }
-
-    public static boolean checkIsCustomEqualBackSlash(Character custom) {
-        return custom == BACK_SLASH;
     }
 }
